rgb effect added

+++ firmware/effect_test/src/Flame.cpp
@@ -4,7 +4,7 @@
 #include <arduino.h>
 #include <FastLED.h>
 
-Flame::Flame(CRGB *leds, uint8_t fps, uint8_t cooling, uint8_t sparking)
+Flame::Flame(CRGB *leds, uint32_t fps, uint8_t cooling, uint8_t sparking)
     : Effect(leds, 1000 / fps)
 {
     this->_cooling = cooling;
